WebCore-777.1
Imported from WebCore-777.1.tar.gz

+++ rendering/RenderTable.h
@@ -74,7 +74,7 @@ public:
         {
         }
 
-        unsigned short span;
+        unsigned span;
         unsigned width; // the calculated position of the column
     };
